+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Remove `Slider` component, use [@goodhood/components Slider](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
 - Remove `Markdown` component, use [@goodhood/components Markdown](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
 - Remove `ContentHeader` component, use [@goodhood/components ContentHeader](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
+- Remove `LinkHeader` component, use [@goodhood/components LinkHeader](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
 
 ## v10.0.0 Breaking changes